neural_network_training/requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.1.0
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
chex==0.1.82
click==8.1.3
colorama==0.4.4
cycler==0.11.0
equinox==0.10.11
filelock==3.9.0
fonttools==4.33.3
frozenlist==1.3.3
fsspec==2023.1.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.3
imageio==2.19.1
importlib-metadata==6.0.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.2
lazy_loader==0.3
lightning-utilities==0.6.0.post0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.2
ml-dtypes==0.2.0
mpmath==1.2.1
multidict==6.0.4
networkx==2.8
nltk==3.8.1
numpy==1.25.2
oauthlib==3.2.2
onnx==1.13.0
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
optax==0.1.7
packaging==21.3
pandas==1.5.3
Pillow==9.1.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pyreadr==0.4.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.11.2
seaborn==0.12.2
seaborn-qqplot==0.5.0
six==1.16.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.5.4
toolz==0.12.0
tqdm==4.64.0
typeguard==4.1.5
typing_extensions==4.7.1
urllib3==1.26.9
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
